added app folder containing cli functionality
